Fikser menu-list part for main-article chapters. Refaktor av menu-list part komponent.
