+++ src/components/parts/menu-list/MenuListPart.less
@@ -1,4 +1,4 @@
-.related-info {
+.menu-list {
     display: flex;
     flex-direction: column;
     width: 100%;
